Seperate outbound router crates
