+++ bridges/snowbridge/primitives/outbound-router/README.md
@@ -1,0 +1,4 @@
+# Outbound Router Primitives
+
+Outbound router logic. Does XCM conversion to a lowered, simpler format the Ethereum contracts can
+understand.
