+++ bridges/snowbridge/primitives/outbound-router/src/lib.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023 Snowfork <hello@snowfork.com>
-// SPDX-FileCopyrightText: 2021-2022 Parity Technologies (UK) Ltd.
+#![cfg_attr(not(feature = "std"), no_std)]
+
 pub mod v1;
 pub mod v2;
