+++ bridges/snowbridge/primitives/router/src/lib.rs
@@ -3,4 +3,3 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod inbound;
-pub mod outbound;
