+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-westend/Cargo.toml
@@ -47,6 +47,7 @@ bridge-hub-westend-runtime = { workspace = true }
 # Snowbridge
 snowbridge-core = { workspace = true }
 snowbridge-router-primitives = { workspace = true }
+snowbridge-outbound-router-primitives = { workspace = true }
 snowbridge-pallet-system = { workspace = true }
 snowbridge-pallet-outbound-queue = { workspace = true }
 snowbridge-pallet-inbound-queue = { workspace = true }
