+++ cumulus/parachains/runtimes/assets/asset-hub-westend/Cargo.toml
@@ -98,6 +98,7 @@ bp-asset-hub-westend = { workspace = true }
 bp-bridge-hub-rococo = { workspace = true }
 bp-bridge-hub-westend = { workspace = true }
 snowbridge-router-primitives = { workspace = true }
+snowbridge-outbound-router-primitives = { workspace = true }
 
 [dev-dependencies]
 asset-test-utils = { workspace = true, default-features = true }
@@ -143,6 +144,7 @@ runtime-benchmarks = [
 	"parachains-common/runtime-benchmarks",
 	"polkadot-parachain-primitives/runtime-benchmarks",
 	"polkadot-runtime-common/runtime-benchmarks",
+	"snowbridge-outbound-router-primitives/runtime-benchmarks",
 	"snowbridge-router-primitives/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 	"xcm-builder/runtime-benchmarks",
@@ -243,6 +245,7 @@ std = [
 	"primitive-types/std",
 	"scale-info/std",
 	"serde_json/std",
+	"snowbridge-outbound-router-primitives/std",
 	"snowbridge-router-primitives/std",
 	"sp-api/std",
 	"sp-block-builder/std",
