+++ cumulus/parachains/runtimes/assets/asset-hub-westend/src/xcm_config.rs
@@ -715,9 +715,9 @@ pub mod bridging {
 		}
 
 		pub type EthereumNetworkExportTableV2 =
-			snowbridge_router_primitives::outbound::v2::XcmFilterExporter<
+			snowbridge_outbound_router_primitives::v2::XcmFilterExporter<
 				xcm_builder::NetworkExportTable<EthereumBridgeTableV2>,
-				snowbridge_router_primitives::outbound::v2::XcmForSnowbridgeV2,
+				snowbridge_outbound_router_primitives::v2::XcmForSnowbridgeV2,
 			>;
 
 		pub type EthereumNetworkExportTable = xcm_builder::NetworkExportTable<EthereumBridgeTable>;
