+++ cumulus/parachains/runtimes/bridge-hubs/bridge-hub-westend/src/bridge_to_ethereum_config.rs
@@ -25,10 +25,10 @@ use crate::{
 use parachains_common::{AccountId, Balance};
 use snowbridge_beacon_primitives::{Fork, ForkVersions};
 use snowbridge_core::{gwei, meth, AllowSiblingsOnly, PricingParameters, Rewards};
-use snowbridge_router_primitives::{
-	inbound::v1::MessageToXcm,
-	outbound::{v1::EthereumBlobExporter, v2::EthereumBlobExporter as EthereumBlobExporterV2},
+use snowbridge_outbound_router_primitives::{
+	v1::EthereumBlobExporter, v2::EthereumBlobExporter as EthereumBlobExporterV2,
 };
+use snowbridge_router_primitives::inbound::v1::MessageToXcm;
 use sp_core::H160;
 use testnet_parachains_constants::westend::{
 	currency::*,
